refactored AnnotationFeatureForm.class from AnnotationDetailEditorPanel.class in its own file
