+++ webanno-ui-annotation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/annotation/detail/AnnotationDetailEditorPanel.java
@@ -49,36 +49,12 @@ import org.apache.uima.fit.util.CasUtil;
 import org.apache.uima.jcas.JCas;
 import org.apache.wicket.Component;
 import org.apache.wicket.ajax.AjaxRequestTarget;
-import org.apache.wicket.ajax.attributes.AjaxCallListener;
-import org.apache.wicket.ajax.attributes.AjaxRequestAttributes;
-import org.apache.wicket.ajax.attributes.IAjaxCallListener;
-import org.apache.wicket.ajax.attributes.ThrottlingSettings;
-import org.apache.wicket.ajax.form.AjaxFormComponentUpdatingBehavior;
 import org.apache.wicket.ajax.form.AjaxFormValidatingBehavior;
-import org.apache.wicket.ajax.markup.html.form.AjaxButton;
-import org.apache.wicket.behavior.AttributeAppender;
-import org.apache.wicket.extensions.ajax.markup.html.modal.ModalWindow;
-import org.apache.wicket.markup.html.WebMarkupContainer;
 import org.apache.wicket.markup.html.basic.Label;
-import org.apache.wicket.markup.html.form.CheckBox;
-import org.apache.wicket.markup.html.form.ChoiceRenderer;
-import org.apache.wicket.markup.html.form.DropDownChoice;
-import org.apache.wicket.markup.html.form.Form;
 import org.apache.wicket.markup.html.panel.FeedbackPanel;
 import org.apache.wicket.markup.html.panel.Panel;
-import org.apache.wicket.markup.repeater.Item;
-import org.apache.wicket.markup.repeater.RefreshingView;
-import org.apache.wicket.markup.repeater.util.ModelIteratorAdapter;
-import org.apache.wicket.model.CompoundPropertyModel;
 import org.apache.wicket.model.IModel;
-import org.apache.wicket.model.Model;
-import org.apache.wicket.model.PropertyModel;
-import org.apache.wicket.request.Request;
-import org.apache.wicket.request.cycle.RequestCycle;
 import org.apache.wicket.spring.injection.annot.SpringBean;
-import org.apache.wicket.util.time.Duration;
-
-import com.googlecode.wicket.kendo.ui.form.TextField;
 
 import de.tudarmstadt.ukp.clarin.webanno.api.AnnotationSchemaService;
 import de.tudarmstadt.ukp.clarin.webanno.api.DocumentService;
@@ -97,8 +73,6 @@ import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.Selection;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.model.VID;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.TypeUtil;
 import de.tudarmstadt.ukp.clarin.webanno.api.annotation.util.WebAnnoCasUtil;
-import de.tudarmstadt.ukp.clarin.webanno.brat.message.SpanAnnotationResponse;
-import de.tudarmstadt.ukp.clarin.webanno.brat.util.JavascriptUtils;
 import de.tudarmstadt.ukp.clarin.webanno.constraints.evaluator.Evaluator;
 import de.tudarmstadt.ukp.clarin.webanno.constraints.evaluator.PossibleValue;
 import de.tudarmstadt.ukp.clarin.webanno.constraints.evaluator.RulesIndicator;
@@ -110,18 +84,8 @@ import de.tudarmstadt.ukp.clarin.webanno.model.Mode;
 import de.tudarmstadt.ukp.clarin.webanno.model.SourceDocumentState;
 import de.tudarmstadt.ukp.clarin.webanno.model.Tag;
 import de.tudarmstadt.ukp.clarin.webanno.model.TagSet;
-import de.tudarmstadt.ukp.clarin.webanno.support.DefaultFocusBehavior;
-import de.tudarmstadt.ukp.clarin.webanno.support.DefaultFocusBehavior2;
-import de.tudarmstadt.ukp.clarin.webanno.support.DescriptionTooltipBehavior;
-import de.tudarmstadt.ukp.clarin.webanno.ui.annotation.component.DeleteOrReplaceAnnotationModalPanel;
-import de.tudarmstadt.ukp.clarin.webanno.ui.annotation.detail.editor.BooleanFeatureEditor;
-import de.tudarmstadt.ukp.clarin.webanno.ui.annotation.detail.editor.FeatureEditor;
-import de.tudarmstadt.ukp.clarin.webanno.ui.annotation.detail.editor.LinkFeatureEditor;
-import de.tudarmstadt.ukp.clarin.webanno.ui.annotation.detail.editor.NumberFeatureEditor;
-import de.tudarmstadt.ukp.clarin.webanno.ui.annotation.detail.editor.TextFeatureEditor;
 import de.tudarmstadt.ukp.dkpro.core.api.lexmorph.type.pos.POS;
 import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Sentence;
-import de.tudarmstadt.ukp.dkpro.core.api.segmentation.type.Token;
 import de.tudarmstadt.ukp.dkpro.core.api.syntax.type.dependency.Dependency;
 
 /**
@@ -156,16 +120,17 @@ public class AnnotationDetailEditorPanel
     public AnnotationDetailEditorPanel(String id, IModel<AnnotatorState> aModel)
     {
         super(id, aModel);
-        
+
         setOutputMarkupId(true);
-        
-        annotationFeatureForm = new AnnotationFeatureForm("annotationFeatureForm", getModel());
+
+        annotationFeatureForm = new AnnotationFeatureForm(this, "annotationFeatureForm",
+            getModel());
         annotationFeatureForm.setOutputMarkupId(true);
-        annotationFeatureForm.add(new AjaxFormValidatingBehavior("submit") { 
+        annotationFeatureForm.add(new AjaxFormValidatingBehavior("submit") {
 			private static final long serialVersionUID = -5642108496844056023L;
 
-			@Override 
-            protected void onSubmit(AjaxRequestTarget aTarget) { 
+			@Override
+            protected void onSubmit(AjaxRequestTarget aTarget) {
                 try {
                     JCas jCas = getEditorCas();
                     actionCreateOrUpdate(aTarget, jCas);
@@ -173,786 +138,20 @@ public class AnnotationDetailEditorPanel
                 catch (Exception e) {
                     handleException(annotationFeatureForm, aTarget, e);
                 }
-            } 
-        }); 
+            }
+        });
         add(annotationFeatureForm);
     }
 
-    private boolean isAnnotationFinished()
+    boolean isAnnotationFinished()
     {
         AnnotatorState state = getModelObject();
-        
+
         if (state.getMode().equals(Mode.CURATION)) {
             return state.getDocument().getState().equals(SourceDocumentState.CURATION_FINISHED);
         }
         else {
             return documentService.isAnnotationFinished(state.getDocument(), state.getUser());
-        }
-    }
-
-    public class AnnotationFeatureForm
-        extends Form<AnnotatorState>
-    {
-        private static final long serialVersionUID = 3635145598405490893L;
-        
-        // Add "featureEditorPanel" to AjaxRequestTargets instead of "featureEditorPanelContent"
-        private WebMarkupContainer featureEditorPanel;
-        private FeatureEditorPanelContent featureEditorPanelContent;
-        
-        private CheckBox forwardAnnotationCheck;
-        private AjaxButton deleteButton;
-        private AjaxButton reverseButton;
-        private LayerSelector layerSelector;   
-        private String selectedTag = "";
-        private Label selectedAnnotationLayer;
-        private TextField<String> forwardAnnotationText;
-        private ModalWindow deleteModal;
-        private Label selectedTextLabel;
-        private List<AnnotationLayer> annotationLayers = new ArrayList<>();
-
-        public AnnotationFeatureForm(String id, IModel<AnnotatorState> aBModel)
-        {
-            super(id, new CompoundPropertyModel<>(aBModel));
-
-            add(forwardAnnotationCheck = new CheckBox("forwardAnnotation")
-            {
-                private static final long serialVersionUID = 8908304272310098353L;
-
-                {
-                    setOutputMarkupId(true);
-                    add(new AjaxFormComponentUpdatingBehavior("change")
-                    {
-                        private static final long serialVersionUID = 5179816588460867471L;
-
-                        @Override
-                        protected void onUpdate(AjaxRequestTarget aTarget)
-                        {
-                            updateForwardAnnotation();
-                            if(AnnotationFeatureForm.this.getModelObject().isForwardAnnotation()){
-                                aTarget.appendJavaScript(JavascriptUtils.getFocusScript(forwardAnnotationText));
-                                selectedTag = "";
-                            }
-                        }
-                    });
-                }
-                
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-
-                    setEnabled(isForwardable());
-                    updateForwardAnnotation();
-                }
-            });
-
-            add(new Label("noAnnotationWarning", "No annotation selected!")
-            {
-                private static final long serialVersionUID = -6046409838139863541L;
-
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-                    setVisible(!AnnotationFeatureForm.this.getModelObject().getSelection()
-                            .getAnnotation().isSet());
-                }
-            });
-
-            add(deleteButton = new AjaxButton("delete")
-            {
-                private static final long serialVersionUID = 1L;
-
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-                    
-                    AnnotatorState state = AnnotationFeatureForm.this.getModelObject();
-                    
-                    setVisible(state.getSelection().getAnnotation().isSet());
-
-                    // Avoid deleting in read-only layers
-                    setEnabled(state.getSelectedAnnotationLayer() != null
-                            && !state.getSelectedAnnotationLayer().isReadonly());
-                }
-
-                @Override
-                public void onSubmit(AjaxRequestTarget aTarget, Form<?> aForm)
-                {
-                    try {
-                        AnnotatorState state = AnnotationFeatureForm.this.getModelObject();
-                        
-                        JCas jCas = getEditorCas();
-                        AnnotationFS fs = selectByAddr(jCas, state.getSelection().getAnnotation().getId());
-
-                        AnnotationLayer layer = state.getSelectedAnnotationLayer();
-                        TypeAdapter adapter = getAdapter(annotationService, layer);
-                        if (adapter instanceof SpanAdapter && getAttachedRels(jCas, fs, layer).size() > 0) {
-                            deleteModal.setTitle("Are you sure you like to delete all attached relations to this span annotation?");
-                            deleteModal.setContent(new DeleteOrReplaceAnnotationModalPanel(
-                                    deleteModal.getContentId(), state, deleteModal,
-                                    AnnotationDetailEditorPanel.this,
-                                    state.getSelectedAnnotationLayer(), false));
-                            deleteModal.show(aTarget);
-                        }
-                        else {
-                            actionDelete(aTarget);
-                        }
-                    }
-                    catch (Exception e) {
-                        handleException(this, aTarget, e);
-                    }
-                }
-            });
-
-            add(reverseButton = new AjaxButton("reverse")
-            {
-                private static final long serialVersionUID = 1L;
-
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-
-                    AnnotatorState state = AnnotationFeatureForm.this.getModelObject();
-                    
-                    setVisible(state.getSelection().getAnnotation().isSet()
-                            && state.getSelection().isArc()
-                            && state.getSelectedAnnotationLayer().getType()
-                                    .equals(WebAnnoConst.RELATION_TYPE));
-
-                    // Avoid reversing in read-only layers
-                    setEnabled(state.getSelectedAnnotationLayer() != null
-                            && !state.getSelectedAnnotationLayer().isReadonly());
-                }
-
-                @Override
-                public void onSubmit(AjaxRequestTarget aTarget, Form<?> aForm)
-                {
-                    aTarget.addChildren(getPage(), FeedbackPanel.class);
-                    try {
-                        actionReverse(aTarget);
-                    }
-                    catch (Exception e) {
-                        handleException(this, aTarget, e);
-                    }
-                }
-            });
-            reverseButton.setOutputMarkupPlaceholderTag(true);
-            
-            add(new AjaxButton("clear")
-            {
-                private static final long serialVersionUID = 1L;
-
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-                    setVisible(AnnotationFeatureForm.this.getModelObject().getSelection()
-                            .getAnnotation().isSet());
-                }
-
-                @Override
-                public void onSubmit(AjaxRequestTarget aTarget, Form<?> aForm)
-                {
-                    aTarget.addChildren(getPage(), FeedbackPanel.class);
-
-                    try {
-                        actionClear(aTarget);
-                    }
-                    catch (Exception e) {
-                        handleException(this, aTarget, e);
-                    }
-                }
-            });
-
-            add(layerSelector = new LayerSelector("defaultAnnotationLayer", annotationLayers));
-
-            featureEditorPanel = new WebMarkupContainer("featureEditorsContainer")
-            {
-                private static final long serialVersionUID = 8908304272310098353L;
-
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-                    setVisible(AnnotationFeatureForm.this.getModelObject().getSelection()
-                            .getAnnotation().isSet());
-                }
-            };
-            // Add placeholder since wmc might start out invisible. Without the placeholder we
-            // cannot make it visible in an AJAX call
-            featureEditorPanel.setOutputMarkupPlaceholderTag(true);
-            featureEditorPanel.setOutputMarkupId(true);
-            
-            featureEditorPanel.add(new Label("noFeaturesWarning", "No features available!") {
-                private static final long serialVersionUID = 4398704672665066763L;
-
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-                    setVisible(AnnotationFeatureForm.this.getModelObject().getFeatureStates()
-                            .isEmpty());
-                }
-            });
-            
-            featureEditorPanelContent = new FeatureEditorPanelContent("featureValues");
-            featureEditorPanel.add(featureEditorPanelContent);
-            
-			forwardAnnotationText = new TextField<>("forwardAnno");
-			forwardAnnotationText.setOutputMarkupId(true);
-			forwardAnnotationText.add(new AjaxFormComponentUpdatingBehavior("keyup") {
-				private static final long serialVersionUID = 4554834769861958396L;
-				
-                @Override
-                protected void updateAjaxAttributes(AjaxRequestAttributes attributes)
-                {
-                    super.updateAjaxAttributes(attributes);
-
-                    IAjaxCallListener listener = new AjaxCallListener()
-                    {
-                        private static final long serialVersionUID = -7968540662654079601L;
-
-                        @Override
-                        public CharSequence getPrecondition(Component component)
-                        {
-                            return "var keycode = Wicket.Event.keyCode(attrs.event);"
-                                    + "    return true;";
-                        }
-                    };
-                    attributes.getAjaxCallListeners().add(listener);
-
-                    attributes.getDynamicExtraParameters()
-                            .add("var eventKeycode = Wicket.Event.keyCode(attrs.event);"
-                                    + "return {keycode: eventKeycode};");
-                    attributes.setAllowDefault(true);
-                }
-				   
-				@Override
-                protected void onUpdate(AjaxRequestTarget aTarget) {	
-					final Request request = RequestCycle.get().getRequest();
-		            final String jsKeycode = request.getRequestParameters()
-		                            .getParameterValue("keycode").toString("");
-		            if (jsKeycode.equals("32")){
-                        try {
-                            JCas jCas = getEditorCas();
-                            actionCreateOrUpdate(aTarget, jCas);
-                            selectedTag = "";
-                        }
-                        catch (Exception e) {
-                            handleException(forwardAnnotationText, aTarget, e);
-                        }
-		            	return;
-		            }
-		            if (jsKeycode.equals("13")){
-		            	selectedTag ="";
-		            	return;
-		            }
-					selectedTag = (forwardAnnotationText.getModelObject() == null ? ""
-							: forwardAnnotationText.getModelObject().charAt(0)) + selectedTag;
-					Map<String, String> bindTags = getBindTags();
-					if (!bindTags.isEmpty()) {
-				        List<FeatureState> featureStates = getModelObject().getFeatureStates();
-					    featureStates.get(0).value = getKeyBindValue(selectedTag, bindTags);
-					}
-					aTarget.add(forwardAnnotationText);
-					aTarget.add(featureEditorPanelContent.get(0));
-				}
-			});
-            forwardAnnotationText.setOutputMarkupId(true);
-            forwardAnnotationText.add(new AttributeAppender("style", "opacity:0", ";"));
-            // forwardAnno.add(new AttributeAppender("style", "filter:alpha(opacity=0)", ";"));
-            add(forwardAnnotationText);
-            
-            // the selected text for annotation
-            selectedTextLabel = new Label("selectedText", PropertyModel.of(getModelObject(),
-                    "selection.text"));
-            selectedTextLabel.setOutputMarkupId(true);
-            featureEditorPanel.add(selectedTextLabel);
-            
-            featureEditorPanel.add(new Label("layerName","Layer"){
-                private static final long serialVersionUID = 6084341323607243784L;
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-                    setVisible(AnnotationFeatureForm.this.getModelObject().getPreferences()
-                            .isRememberLayer());
-                }
-                
-            });
-            featureEditorPanel.setOutputMarkupId(true);
-
-            // the annotation layer for the selected annotation
-            selectedAnnotationLayer = new Label("selectedAnnotationLayer", new Model<String>())
-            {
-                private static final long serialVersionUID = 4059460390544343324L;
-
-                @Override
-                protected void onConfigure()
-                {
-                    super.onConfigure();
-                    setVisible(AnnotationFeatureForm.this.getModelObject().getPreferences()
-                            .isRememberLayer());
-                }
-
-            };
-            selectedAnnotationLayer.setOutputMarkupId(true);
-            featureEditorPanel.add(selectedAnnotationLayer);
-            
-            add(featureEditorPanel);
-            
-            add(deleteModal = new ModalWindow("yesNoModal"));
-            deleteModal.setOutputMarkupId(true);
-
-            deleteModal.setInitialWidth(600);
-            deleteModal.setInitialHeight(50);
-            deleteModal.setResizable(true);
-            deleteModal.setWidthUnit("px");
-            deleteModal.setHeightUnit("px");
-            deleteModal.setTitle("Are you sure you want to delete the existing annotation?");
-        }
-        
-        @Override
-        protected void onConfigure()
-        {
-            super.onConfigure();
-
-            // Avoid reversing in read-only layers
-            setEnabled(getModelObject().getDocument() != null && !isAnnotationFinished());
-        }
-        
-        private String getKeyBindValue(String aKey, Map<String, String> aBindTags){
-            // check if all the key pressed are the same character
-            // if not, just check a Tag for the last char pressed
-            if(aKey.isEmpty()){
-                return aBindTags.get(aBindTags.keySet().iterator().next());
-            }
-            char prevC = aKey.charAt(0);
-            for(char ch:aKey.toCharArray()){
-                if(ch!=prevC){
-                    break;
-                }
-            }
-            
-            if (aBindTags.get(aKey)!=null){
-                return aBindTags.get(aKey);
-            }
-            // re-cycle suggestions
-            if(aBindTags.containsKey(aKey.substring(0,1))){
-                selectedTag = aKey.substring(0,1);
-                return aBindTags.get(aKey.substring(0,1));
-            }
-            // set it to the first in the tag list , when arbitrary key is pressed
-            return aBindTags.get(aBindTags.keySet().iterator().next());
-        }
-        
-        private Map<String, String> getBindTags()
-        {
-            AnnotationFeature f = annotationService
-                    .listAnnotationFeature(getModelObject().getSelectedAnnotationLayer()).get(0);
-            TagSet tagSet = f.getTagset();
-            Map<Character, String> tagNames = new LinkedHashMap<>();
-            Map<String, String> bindTag2Key = new LinkedHashMap<>();
-            for (Tag tag : annotationService.listTags(tagSet)) {
-                if (tagNames.containsKey(tag.getName().toLowerCase().charAt(0))) {
-                    String oldBinding = tagNames.get(tag.getName().toLowerCase().charAt(0));
-                    String newBinding = oldBinding + tag.getName().toLowerCase().charAt(0);
-                    tagNames.put(tag.getName().toLowerCase().charAt(0), newBinding);
-                    bindTag2Key.put(newBinding, tag.getName());
-                }
-                else {
-                    tagNames.put(tag.getName().toLowerCase().charAt(0),
-                            tag.getName().toLowerCase().substring(0, 1));
-                    bindTag2Key.put(tag.getName().toLowerCase().substring(0, 1), tag.getName());
-                }
-            }
-            return bindTag2Key;
-
-        }
-        
-        private void updateForwardAnnotation()
-        {
-            AnnotatorState state = getModelObject();
-
-            if (state.getSelectedAnnotationLayer() != null
-                    && !state.getSelectedAnnotationLayer().isLockToTokenOffset()) {
-                state.setForwardAnnotation(false);// no forwarding for
-                                                  // sub-/multitoken annotation
-            }
-            else {
-                state.setForwardAnnotation(state.isForwardAnnotation());
-            }
-        }
-        
-        private void updateLayersDropdown()
-        {
-            LOG.trace("updateLayersDropdown()");
-            
-            AnnotatorState state = getModelObject();
-            
-            annotationLayers.clear();
-            AnnotationLayer l = null;
-            for (AnnotationLayer layer : state.getAnnotationLayers()) {
-                if (!layer.isEnabled() || layer.isReadonly()
-                        || layer.getName().equals(Token.class.getName())) {
-                    continue;
-                }
-                if (layer.getType().equals(WebAnnoConst.SPAN_TYPE)) {
-                    annotationLayers.add(layer);
-                    l = layer;
-                }
-                // manage chain type
-                else if (layer.getType().equals(WebAnnoConst.CHAIN_TYPE)) {
-                    for (AnnotationFeature feature : annotationService.listAnnotationFeature(layer)) {
-                        if (!feature.isEnabled()) {
-                            continue;
-                        }
-                        if (feature.getName().equals(WebAnnoConst.COREFERENCE_TYPE_FEATURE)) {
-                            annotationLayers.add(layer);
-                        }
-
-                    }
-                }
-                // chain
-            }
-            if (state.getDefaultAnnotationLayer() != null) {
-                state.setSelectedAnnotationLayer(state.getDefaultAnnotationLayer());
-            }
-            else if (l != null) {
-                state.setSelectedAnnotationLayer(l);
-            }
-        }
-        
-        private void updateRememberLayer()
-        {
-            LOG.trace("updateRememberLayer()");
-            
-            AnnotatorState state = getModelObject();
-            
-            if (state.getPreferences().isRememberLayer()) {
-                if (state.getDefaultAnnotationLayer() == null) {
-                    state.setDefaultAnnotationLayer(state.getSelectedAnnotationLayer());
-                }
-            } 
-            else if (!state.getSelection().isArc()) {
-                state.setDefaultAnnotationLayer(state.getSelectedAnnotationLayer());
-            }
-            
-            // if no layer is selected in Settings
-            if (state.getSelectedAnnotationLayer() != null) {
-                selectedAnnotationLayer.setDefaultModelObject(
-                        state.getSelectedAnnotationLayer().getUiName());
-            } 
-        }
-        
-        public class LayerSelector
-            extends DropDownChoice<AnnotationLayer>
-        {
-            private static final long serialVersionUID = 2233133653137312264L;
-
-            public LayerSelector(String aId, List<? extends AnnotationLayer> aChoices)
-            {
-                super(aId, aChoices);
-                setOutputMarkupId(true);
-                setChoiceRenderer(new ChoiceRenderer<>("uiName"));
-                add(new AjaxFormComponentUpdatingBehavior("change")
-                {
-                    private static final long serialVersionUID = 5179816588460867471L;
-
-                    @Override
-                    protected void onUpdate(AjaxRequestTarget aTarget)
-                    {
-                        AnnotatorState state = AnnotationFeatureForm.this.getModelObject();
-                        
-                        // If "remember layer" is set, the we really just update the selected
-                        // layer...
-                        // we do not touch the selected annotation not the annotation detail panel
-                        if (state.getPreferences().isRememberLayer()) {
-                            state.setSelectedAnnotationLayer(getModelObject());
-                        }
-                        // If "remember layer" is not set, then changing the layer means that we
-                        // want to change the type of the currently selected annotation
-                        else if (!state.getSelectedAnnotationLayer().equals(getModelObject())
-                                && state.getSelection().getAnnotation().isSet()) {
-                            if (state.getSelection().isArc()) {
-                                try {
-                                    actionClear(aTarget);
-                                }
-                                catch (Exception e) {
-                                    handleException(LayerSelector.this, aTarget, e);
-                                }
-                            }
-                            else {
-                                AnnotationFeatureForm.this.deleteModal
-                                        .setContent(new DeleteOrReplaceAnnotationModalPanel(
-                                                AnnotationFeatureForm.this.deleteModal.getContentId(),
-                                                state, AnnotationFeatureForm.this.deleteModal,
-                                                AnnotationDetailEditorPanel.this, getModelObject(),
-                                                true));
-
-                                AnnotationFeatureForm.this.deleteModal.setWindowClosedCallback(
-                                        new ModalWindow.WindowClosedCallback()
-                                        {
-                                            private static final long serialVersionUID = 4364820331676014559L;
-
-                                            @Override
-                                            public void onClose(AjaxRequestTarget target)
-                                            {
-                                                target.add(AnnotationFeatureForm.this);
-                                            }
-                                        });
-                                AnnotationFeatureForm.this.deleteModal.show(aTarget);
-                            }
-                        }
-                        // If no annotation is selected, then prime the annotation detail panel for
-                        // the new type
-                        else {
-                            state.setSelectedAnnotationLayer(getModelObject());
-                            AnnotationFeatureForm.this.selectedAnnotationLayer
-                                    .setDefaultModelObject(getModelObject().getUiName());
-                            aTarget.add(AnnotationFeatureForm.this.selectedAnnotationLayer);
-                            clearFeatureEditorModels(aTarget);
-                        }
-                    }
-                });
-            }
-        }
-        
-        public class FeatureEditorPanelContent
-            extends RefreshingView<FeatureState>
-        {
-            private static final long serialVersionUID = -8359786805333207043L;
-
-            public FeatureEditorPanelContent(String aId)
-            {
-                super(aId);
-                setOutputMarkupId(true);
-                // This strategy caches items as long as the panel exists. This is important to
-                // allow the Kendo ComboBox datasources to be re-read when constraints change the
-                // available tags.
-                setItemReuseStrategy(new CachingReuseStrategy());
-            }
-
-            @SuppressWarnings("rawtypes")
-            @Override
-            protected void populateItem(final Item<FeatureState> item)
-            {
-                LOG.trace("FeatureEditorPanelContent.populateItem("
-                        + item.getModelObject().feature.getUiName() + ": "
-                        + item.getModelObject().value + ")");
-                
-                // Feature editors that allow multiple values may want to update themselves,
-                // e.g. to add another slot.
-                item.setOutputMarkupId(true);
-
-                final FeatureState featureState = item.getModelObject();
-
-                final FeatureEditor frag;
-                switch (featureState.feature.getMultiValueMode()) {
-                case NONE: {
-                    switch (featureState.feature.getType()) {
-                    case CAS.TYPE_NAME_INTEGER: {
-                        frag = new NumberFeatureEditor("editor", "numberFeatureEditor",
-                                AnnotationFeatureForm.this, item.getModel());
-                        break;
-                    }
-                    case CAS.TYPE_NAME_FLOAT: {
-                        frag = new NumberFeatureEditor("editor", "numberFeatureEditor",
-                                AnnotationFeatureForm.this, item.getModel());
-                        break;
-                    }
-                    case CAS.TYPE_NAME_BOOLEAN: {
-                        frag = new BooleanFeatureEditor("editor", "booleanFeatureEditor",
-                                AnnotationFeatureForm.this, item.getModel());
-                        break;
-                    }
-                    case CAS.TYPE_NAME_STRING: {
-                        frag = new TextFeatureEditor("editor", "textFeatureEditor",
-                                AnnotationFeatureForm.this, item.getModel());
-                        break;
-                    }
-                    default:
-                        throw new IllegalArgumentException(
-                            "Unsupported type [" + featureState.feature.getType()
-                                    + "] on feature [" + featureState.feature.getName() + "]");
-                    }
-                    break;
-                }
-                case ARRAY: {
-                    switch (featureState.feature.getLinkMode()) {
-                    case WITH_ROLE: {
-                        // If it is none of the primitive types, it must be a link feature
-                        frag = new LinkFeatureEditor("editor", "linkFeatureEditor",
-                                AnnotationDetailEditorPanel.this, item.getModel());
-                        break;
-
-                    }
-                    default:
-                    throw new IllegalArgumentException(
-                            "Unsupported link mode [" + featureState.feature.getLinkMode()
-                                    + "] on feature [" + featureState.feature.getName() + "]");
-                    }
-                    break;
-                }
-                default:
-                throw new IllegalArgumentException("Unsupported multi-value mode ["
-                        + featureState.feature.getMultiValueMode() + "] on feature ["
-                        + featureState.feature.getName() + "]");
-                }
-
-                if (!featureState.feature.getLayer().isReadonly()) {
-                    AnnotatorState state = AnnotationFeatureForm.this.getModelObject();
-                    
-                    // Whenever it is updating an annotation, it updates automatically when a
-                    // component for the feature lost focus - but updating is for every component
-                    // edited LinkFeatureEditors must be excluded because the auto-update will break
-                    // the ability to add slots. Adding a slot is NOT an annotation action.
-                    if (state.getSelection().getAnnotation().isSet()
-                            && !(frag instanceof LinkFeatureEditor)) {
-                        addAnnotateActionBehavior(frag, "change");
-                    }
-                    else if (!(frag instanceof LinkFeatureEditor)) {
-                        addRefreshFeaturePanelBehavior(frag, "change");
-                    }
-
-                    // Put focus on hidden input field if we are in forward-mode
-                    if (state.isForwardAnnotation()) {
-                        AnnotationFeatureForm.this.forwardAnnotationText
-                                .add(new DefaultFocusBehavior2());
-                    }
-                    // Put focus on first component if we select an existing annotation or create a
-                    // new one
-                    else if (item.getIndex() == 0
-                            && SpanAnnotationResponse.is(state.getAction().getUserAction())) {
-                        frag.getFocusComponent().add(new DefaultFocusBehavior());
-                    }
-                    // Restore/preserve focus when tabbing through the feature editors
-                    else if (state.getAction().getUserAction() == null) {
-                        AjaxRequestTarget target = RequestCycle.get()
-                                .find(AjaxRequestTarget.class);
-                        if (target != null && frag.getFocusComponent().getMarkupId()
-                                .equals(target.getLastFocusedElementId())) {
-                            target.focusComponent(frag.getFocusComponent());
-                        }
-                    }
-
-                    // Add tooltip on label
-                    StringBuilder tooltipTitle = new StringBuilder();
-                    tooltipTitle.append(featureState.feature.getUiName());
-                    if (featureState.feature.getTagset() != null) {
-                        tooltipTitle.append(" (");
-                        tooltipTitle.append(featureState.feature.getTagset().getName());
-                        tooltipTitle.append(')');
-                    }
-
-                    Component labelComponent = frag.getLabelComponent();
-                    labelComponent.add(new AttributeAppender("style", "cursor: help", ";"));
-                    labelComponent.add(new DescriptionTooltipBehavior(tooltipTitle.toString(),
-                            featureState.feature.getDescription()));
-                }
-                else {
-                    frag.getFocusComponent().setEnabled(false);
-                }
-
-                // We need to enable the markup ID here because we use it during the AJAX behavior
-                // that
-                // automatically saves feature editors on change/blur. Check
-                // addAnnotateActionBehavior.
-                frag.setOutputMarkupId(true);
-                item.add(frag);
-            }
-
-            private void addRefreshFeaturePanelBehavior(final FeatureEditor aFrag, String aEvent)
-            {
-                aFrag.getFocusComponent().add(new AjaxFormComponentUpdatingBehavior(aEvent)
-                {
-                    private static final long serialVersionUID = 5179816588460867471L;
-
-                    @Override
-                    protected void onUpdate(AjaxRequestTarget aTarget)
-                    {
-                        aTarget.add(featureEditorPanel);
-                    }
-                });
-            }
-
-            private void addAnnotateActionBehavior(final FeatureEditor aFrag, String aEvent)
-            {
-                aFrag.getFocusComponent().add(new AjaxFormComponentUpdatingBehavior(aEvent)
-                {
-                    private static final long serialVersionUID = 5179816588460867471L;
-
-                    @Override
-                    protected void updateAjaxAttributes(AjaxRequestAttributes aAttributes)
-                    {
-                        super.updateAjaxAttributes(aAttributes);
-                        // When focus is on a feature editor and the user selects a new annotation,
-                        // there is a race condition between the saving the value of the feature
-                        // editor and the loading of the new annotation. Delay the feature editor
-                        // save to give preference to loading the new annotation.
-                        aAttributes.setThrottlingSettings(new ThrottlingSettings(getMarkupId(),
-                                Duration.milliseconds(250), true));
-                        aAttributes.getAjaxCallListeners().add(new AjaxCallListener()
-                        {
-                            private static final long serialVersionUID = 1L;
-
-                            @Override
-                            public CharSequence getPrecondition(Component aComponent)
-                            {
-                                // If the panel refreshes because the user selects a new annotation,
-                                // the annotation editor panel is updated for the new annotation
-                                // first (before saving values) because of the delay set above. When
-                                // the delay is over, we can no longer save the value because the
-                                // old component is no longer there. We use the markup id of the
-                                // editor fragments to check if the old component is still there
-                                // (i.e. if the user has just tabbed to a new field) or if the old
-                                // component is gone (i.e. the user selected/created another
-                                // annotation). If the old component is no longer there, we abort
-                                // the delayed save action.
-                                return "return $('#" + aFrag.getMarkupId() + "').length > 0;";
-                            }
-                        });
-                    }
-
-                    @Override
-                    protected void onUpdate(AjaxRequestTarget aTarget)
-                    {
-                        try {
-                            AnnotatorState state = AnnotationFeatureForm.this.getModelObject();
-                            if (state.getConstraints() != null) {
-                                // Make sure we update the feature editor panel because due to
-                                // constraints the contents may have to be re-rendered
-                                aTarget.add(AnnotationFeatureForm.this.featureEditorPanel);
-                            }
-                            JCas jCas = getEditorCas();
-                            actionCreateOrUpdate(aTarget, jCas);
-                        }
-                        catch (Exception e) {
-                            handleException(FeatureEditorPanelContent.this, aTarget, e);
-                        }
-                    }
-                });
-            }
-
-            @Override
-            protected Iterator<IModel<FeatureState>> getItemModels()
-            {
-                List<FeatureState> featureStates = AnnotationFeatureForm.this.getModelObject()
-                        .getFeatureStates();
-
-                return new ModelIteratorAdapter<FeatureState>(
-                        featureStates)
-                {
-                    @Override
-                    protected IModel<FeatureState> model(FeatureState aObject)
-                    {
-                        return FeatureStateModel.of(AnnotationFeatureForm.this.getModel(), aObject);
-                    }
-                };
-            }
         }
     }
 
@@ -991,25 +190,25 @@ public class AnnotationDetailEditorPanel
             }
         }
     }
-    
+
     private void createNewRelationAnnotation(AjaxRequestTarget aTarget, ArcAdapter aAdapter,
             JCas aJCas)
         throws AnnotationException
     {
         LOG.trace("createNewRelationAnnotation()");
-        
+
         AnnotatorState state = getModelObject();
         Selection selection = state.getSelection();
 
         AnnotationFS originFs = selectByAddr(aJCas, selection.getOrigin());
         AnnotationFS targetFs = selectByAddr(aJCas, selection.getTarget());
-        
+
         // Creating a relation
         AnnotationFS arc = aAdapter.add(originFs, targetFs, aJCas, state.getWindowBeginOffset(),
                 state.getWindowEndOffset(), null, null);
         selection.selectArc(new VID(arc), originFs, targetFs);
     }
-    
+
     private void createNewSpanAnnotation(AjaxRequestTarget aTarget, SpanAdapter aAdapter,
             JCas aJCas)
         throws UIMAException, ClassNotFoundException, IOException, AnnotationException
@@ -1019,7 +218,7 @@ public class AnnotationDetailEditorPanel
         AnnotatorState state = getModelObject();
         Selection selection = state.getSelection();
         List<FeatureState> featureStates = state.getFeatureStates();
-        
+
         for (FeatureState featureState : featureStates) {
             Serializable spanValue = aAdapter.getSpan(aJCas, selection.getBegin(),
                     selection.getEnd(), featureState.feature, null);
@@ -1028,9 +227,10 @@ public class AnnotationDetailEditorPanel
                 if (state.isForwardAnnotation()) {
                     featureState.value = spanValue;
                     featureStates.get(0).value = spanValue;
-                    annotationFeatureForm.selectedTag = annotationFeatureForm.getBindTags()
-                            .entrySet().stream().filter(e -> e.getValue().equals(spanValue))
-                            .map(Map.Entry::getKey).findFirst().orElse(null);
+                    String selectedTag = annotationFeatureForm.getBindTags()
+                        .entrySet().stream().filter(e -> e.getValue().equals(spanValue))
+                        .map(Map.Entry::getKey).findFirst().orElse(null);
+                    annotationFeatureForm.setSelectedTag(selectedTag);
                 }
                 else {
                     actionClear(aTarget);
@@ -1044,17 +244,17 @@ public class AnnotationDetailEditorPanel
         AnnotationFS annoFs = WebAnnoCasUtil.selectByAddr(aJCas, annoId);
         selection.selectSpan(new VID(annoId), aJCas, annoFs.getBegin(), annoFs.getEnd());
     }
-    
+
     private void createNewChainElement(AjaxRequestTarget aTarget, ChainAdapter aAdapter,
             JCas aJCas)
         throws AnnotationException
     {
         LOG.trace("createNewChainElement()");
-        
+
         AnnotatorState state = getModelObject();
         Selection selection = state.getSelection();
         List<FeatureState> featureStates = state.getFeatureStates();
-        
+
         for (FeatureState featureState : featureStates) {
             Serializable spanValue = aAdapter.getSpan(aJCas,
                     selection.getBegin(), selection.getEnd(), featureState.feature, null);
@@ -1063,9 +263,10 @@ public class AnnotationDetailEditorPanel
                 if (state.isForwardAnnotation()) {
                     featureState.value = spanValue;
                     featureStates.get(0).value = spanValue;
-                    annotationFeatureForm.selectedTag = annotationFeatureForm.getBindTags()
-                            .entrySet().stream().filter(e -> e.getValue().equals(spanValue))
-                            .map(Map.Entry::getKey).findFirst().orElse(null);
+                    String selectedTag = annotationFeatureForm.getBindTags()
+                        .entrySet().stream().filter(e -> e.getValue().equals(spanValue))
+                        .map(Map.Entry::getKey).findFirst().orElse(null);
+                    annotationFeatureForm.setSelectedTag(selectedTag);
                 }
             }
         }
@@ -1074,22 +275,22 @@ public class AnnotationDetailEditorPanel
         selection.setText(
                 aJCas.getDocumentText().substring(selection.getBegin(), selection.getEnd()));
     }
-    
+
     private void createNewChainLinkAnnotation(AjaxRequestTarget aTarget, ChainAdapter aAdapter,
             JCas aJCas) {
         LOG.trace("createNewChainLinkAnnotation()");
-        
+
         AnnotatorState state = getModelObject();
         Selection selection = state.getSelection();
-        
+
         AnnotationFS originFs = selectByAddr(aJCas, selection.getOrigin());
         AnnotationFS targetFs = selectByAddr(aJCas, selection.getTarget());
-        
+
         // Creating a new chain link
         int addr = aAdapter.addArc(aJCas, originFs, targetFs, null, null);
         selection.selectArc(new VID(addr), originFs, targetFs);
     }
-    
+
     @Override
     public void actionFillSlot(AjaxRequestTarget aTarget, JCas aJCas, int aBegin, int aEnd,
             VID aVID)
@@ -1103,7 +304,7 @@ public class AnnotationDetailEditorPanel
         if (!state.isSlotArmed()) {
             throw new IllegalStateException("No slot is armed.");
         }
-        
+
         // Fill slot with new annotation (only works if a concrete type is set for the link feature!
         int id;
         if (aVID.isNotSet()) {
@@ -1117,12 +318,12 @@ public class AnnotationDetailEditorPanel
                 throw new AnnotationException(
                         "Unable to create annotation of type [" + CAS.TYPE_NAME_ANNOTATION
                                 + "]. Please click an annotation in stead of selecting new text.");
-            } 
+            }
         }
         else {
             id = aVID.getId();
         }
-        
+
         // Fill the annotation into the slow
         try {
             setSlot(aTarget, aJCas, id);
@@ -1131,7 +332,7 @@ public class AnnotationDetailEditorPanel
             handleException(this, aTarget, e);
         }
     }
-    
+
     @Override
     public void actionSelect(AjaxRequestTarget aTarget, JCas aJCas)
         throws AnnotationException
@@ -1154,20 +355,20 @@ public class AnnotationDetailEditorPanel
         throws UIMAException, ClassNotFoundException, IOException, AnnotationException
     {
         LOG.trace("actionAnnotate(isForwarded: {})", aIsForwarded);
-    
+
         if (isAnnotationFinished()) {
             throw new AnnotationException("This document is already closed. Please ask your "
                     + "project manager to re-open it via the Monitoring page");
         }
-        
+
         AnnotatorState state = getModelObject();
         state.getAction().setAnnotate(true);
-        
+
         // Note that refresh changes the selected layer if a relation is created. Then the layer
         // switches from the selected span layer to the relation layer that is attached to the span
         if (state.getSelection().isArc()) {
             LOG.trace("actionAnnotate() relation annotation - looking for attached layer");
-    
+
             // FIXME REC I think this whole section which meddles around with the selected annotation
             // layer should be moved out of there to the place where we originally set the annotation
             // layer...!
@@ -1175,28 +376,28 @@ public class AnnotationDetailEditorPanel
             AnnotationLayer spanLayer = TypeUtil.getLayer(annotationService, state.getProject(),
                     originFS);
             if (
-                    state.getPreferences().isRememberLayer() && 
-                    state.getAction().isAnnotate() && 
-                    !spanLayer.equals(state.getDefaultAnnotationLayer())) 
+                    state.getPreferences().isRememberLayer() &&
+                    state.getAction().isAnnotate() &&
+                    !spanLayer.equals(state.getDefaultAnnotationLayer()))
             {
                 throw new AnnotationException(
                         "No relation annotation allowed ["+ spanLayer.getUiName() +"]");
             }
-    
+
             AnnotationLayer previousLayer = state.getSelectedAnnotationLayer();
-            
+
             // If we are creating a relation annotation, we have to set the current layer depending
             // on the type of relation that is permitted between the source/target span. This is
             // necessary because we have no separate UI control to set the relation annotation type.
             // It is possible because currently only a single relation layer is allowed to attach to
             // any given span layer.
-            
+
             // If we drag an arc between POS annotations, then the relation must be a dependency
             // relation.
             // FIXME - Actually this case should be covered by the last case - the database lookup!
             if (
-                    spanLayer.isBuiltIn() && 
-                    spanLayer.getName().equals(POS.class.getName())) 
+                    spanLayer.isBuiltIn() &&
+                    spanLayer.getName().equals(POS.class.getName()))
             {
                 AnnotationLayer depLayer = annotationService.getLayer(Dependency.class.getName(),
                         state.getProject());
@@ -1228,32 +429,32 @@ public class AnnotationDetailEditorPanel
                     }
                 }
             }
-    
+
             state.setDefaultAnnotationLayer(spanLayer);
-         
+
             // If we switched layers, we need to initialize the feature editors for the new layer
             if (!Objects.equals(previousLayer, state.getSelectedAnnotationLayer())) {
                 loadFeatureEditorModels(aJCas, aTarget);
             }
         }
-    
+
         LOG.trace("actionAnnotate() selectedLayer: {}",
                 state.getSelectedAnnotationLayer().getUiName());
         LOG.trace("actionAnnotate() defaultLayer: {}",
                 state.getDefaultAnnotationLayer().getUiName());
-        
+
         if (state.getSelectedAnnotationLayer() == null) {
             error("No layer is selected. First select a layer.");
             aTarget.addChildren(getPage(), FeedbackPanel.class);
             return;
         }
-    
+
         if (state.getSelectedAnnotationLayer().isReadonly()) {
             error("Layer is not editable.");
             aTarget.addChildren(getPage(), FeedbackPanel.class);
             return;
         }
-        
+
         // Verify if input is valid according to tagset
         LOG.trace("actionAnnotate() verifying feature values in editors");
         List<FeatureState> featureStates = getModelObject().getFeatureStates();
@@ -1275,51 +476,51 @@ public class AnnotationDetailEditorPanel
                 }
             }
         }
-        
+
         // #186 - After filling a slot, the annotation detail panel is not updated 
-        aTarget.add(annotationFeatureForm.featureEditorPanel);
-    
+        aTarget.add(annotationFeatureForm.getFeatureEditorPanel());
+
         TypeAdapter adapter = getAdapter(annotationService, state.getSelectedAnnotationLayer());
-        
+
         // If this is an annotation creation action, create the annotation
     	if (state.getSelection().getAnnotation().isNotSet()) {
     	    // Load the feature editors with the remembered values (if any)
     	    loadFeatureEditorModels(aJCas, aTarget);
     	    createNewAnnotation(aTarget, adapter, aJCas);
     	}
-    
+
         // Update the features of the selected annotation from the values presently in the
     	// feature editors
     	writeFeatureEditorModelsToCas(adapter, aJCas);
-    
+
         // Update progress information
         LOG.trace("actionAnnotate() updating progress information");
         int sentenceNumber = getSentenceNumber(aJCas, state.getSelection().getBegin());
         state.setFocusUnitIndex(sentenceNumber);
         state.getDocument().setSentenceAccessed(sentenceNumber);
-    
+
         // persist changes
         writeEditorCas(aJCas);
-    
+
         // Remember the current feature values independently for spans and relations
         LOG.trace("actionAnnotate() remembering feature editor values");
         state.rememberFeatures();
-    
+
     	// Loading feature editor values from CAS
     	loadFeatureEditorModels(aJCas, aTarget);
-    	
+
     	// onAnnotate callback
         LOG.trace("onAnnotate()");
     	onAnnotate(aTarget);
-    
+
     	// Handle auto-forward if it is enabled
     	if (state.isForwardAnnotation() && !aIsForwarded && featureStates.get(0).value != null) {
     		if (state.getSelection().getEnd() >= state.getFirstVisibleUnitEnd()) {
     			autoScroll(aJCas, true);
     		}
-    
+
             LOG.info("BEGIN auto-forward annotation");
-    
+
             AnnotationFS nextToken = WebAnnoCasUtil.getNextToken(aJCas, state.getSelection().getBegin(),
                     state.getSelection().getEnd());
             if (nextToken != null) {
@@ -1328,40 +529,41 @@ public class AnnotationDetailEditorPanel
                     actionCreateOrUpdate(aTarget, aJCas, true);
                 }
             }
-    
+
             LOG.trace("onAutoForward()");
     		onAutoForward(aTarget);
-    		
+
             LOG.info("END auto-forward annotation");
-    	} 
+    	}
     	// Perform auto-scroll if it is enabled
     	else if (state.getPreferences().isScrollPage()) {
     		autoScroll(aJCas, false);
     	}
-    	
-    	annotationFeatureForm.forwardAnnotationText.setModelObject(null);
-    	
+
+    	annotationFeatureForm.getForwardAnnotationText().setModelObject(null);
+
         LOG.trace("onChange()");
     	onChange(aTarget);
-    	
+
     	if (state.isForwardAnnotation() && state.getFeatureStates().get(0).value != null) {
     		aTarget.add(annotationFeatureForm);
     	}
-    	
+
         // If we created a new annotation, then refresh the available annotation layers in the
         // detail panel.
         if (state.getSelection().getAnnotation().isNotSet()) {
             // This already happens in loadFeatureEditorModels() above - probably not needed
             // here again
             // annotationFeatureForm.updateLayersDropdown();
-            
+
             LOG.trace("actionAnnotate() setting selected layer (not sure why)");
-            if (annotationFeatureForm.annotationLayers.isEmpty()) {
+            if (annotationFeatureForm.getAnnotationLayers().isEmpty()) {
                 state.setSelectedAnnotationLayer(new AnnotationLayer());
             }
             else if (state.getSelectedAnnotationLayer() == null) {
                 if (state.getRememberedSpanLayer() == null) {
-                    state.setSelectedAnnotationLayer(annotationFeatureForm.annotationLayers.get(0));
+                    state.setSelectedAnnotationLayer(annotationFeatureForm.getAnnotationLayers()
+                        .get(0));
                 }
                 else {
                     state.setSelectedAnnotationLayer(state.getRememberedSpanLayer());
@@ -1369,11 +571,11 @@ public class AnnotationDetailEditorPanel
             }
             LOG.trace("actionAnnotate() selectedLayer: {}",
                     state.getSelectedAnnotationLayer().getUiName());
-            
+
             // Actually not sure why we would want to clear these here - in fact, they should
             // still be around for the rendering phase of the feature editors...
             //clearFeatureEditorModels(aTarget);
-            
+
             // This already happens in loadFeatureEditorModels() above - probably not needed
             // here again
             // annotationFeatureForm.updateRememberLayer();
@@ -1386,9 +588,9 @@ public class AnnotationDetailEditorPanel
         AnnotationException
     {
         JCas jCas = getEditorCas();
-        
+
         AnnotatorState state = getModelObject();
-        
+
         AnnotationFS fs = selectByAddr(jCas, state.getSelection().getAnnotation().getId());
 
         // TODO We assume here that the selected annotation layer corresponds to the type of the
@@ -1510,7 +712,7 @@ public class AnnotationDetailEditorPanel
         JCas jCas = getEditorCas();
 
         AnnotatorState state = getModelObject();
-        
+
         AnnotationFS idFs = selectByAddr(jCas, state.getSelection().getAnnotation().getId());
 
         jCas.removeFsFromIndexes(idFs);
@@ -1519,7 +721,7 @@ public class AnnotationDetailEditorPanel
         AnnotationFS targetFs = selectByAddr(jCas, state.getSelection().getTarget());
 
         List<FeatureState> featureStates = getModelObject().getFeatureStates();
-        
+
         TypeAdapter adapter = getAdapter(annotationService, state.getSelectedAnnotationLayer());
         if (adapter instanceof ArcAdapter) {
             if (featureStates.isEmpty()) {
@@ -1569,7 +771,7 @@ public class AnnotationDetailEditorPanel
         aTarget.add(annotationFeatureForm);
         onChange(aTarget);
     }
-    
+
     public JCas getEditorCas()
         throws IOException {
         AnnotatorState state = getModelObject();
@@ -1583,7 +785,7 @@ public class AnnotationDetailEditorPanel
             return curationDocumentService.readCurationCas(state.getDocument());
         }
     }
-    
+
     public void writeEditorCas(JCas aJCas)
         throws IOException
     {
@@ -1606,7 +808,7 @@ public class AnnotationDetailEditorPanel
     private void autoScroll(JCas jCas, boolean aForward)
     {
         AnnotatorState state = getModelObject();
-        
+
         if (aForward) {
             // Fetch the first sentence on screen
             Sentence sentence = selectByAddr(jCas, Sentence.class,
@@ -1634,7 +836,7 @@ public class AnnotationDetailEditorPanel
     public void setSlot(AjaxRequestTarget aTarget, JCas aJCas, int aAnnotationId)
     {
         AnnotatorState state = getModelObject();
-        
+
         // Set an armed slot
         if (!state.getSelection().isArc() && state.isSlotArmed()) {
             List<LinkWithRoleModel> links = (List<LinkWithRoleModel>) state.getFeatureState(state
@@ -1653,11 +855,11 @@ public class AnnotationDetailEditorPanel
                 handleException(this, aTarget, e);
             }
         }
-        
+
         state.clearArmedSlot();
     }
-    
-    public void loadFeatureEditorModels(AjaxRequestTarget aTarget) 
+
+    public void loadFeatureEditorModels(AjaxRequestTarget aTarget)
         throws AnnotationException
     {
         try {
@@ -1671,12 +873,12 @@ public class AnnotationDetailEditorPanel
             throw new AnnotationException(e);
         }
     }
-    
+
     public void loadFeatureEditorModels(JCas aJCas, AjaxRequestTarget aTarget)
         throws AnnotationException
     {
         LOG.trace("loadFeatureEditorModels()");
-        
+
         AnnotatorState state = getModelObject();
         Selection selection = state.getSelection();
 
@@ -1686,18 +888,18 @@ public class AnnotationDetailEditorPanel
                 featureState.value = new ArrayList<>();
             }
         }
-        
+
         try {
             if (selection.isSpan()) {
                 annotationFeatureForm.updateLayersDropdown();
             }
-            
+
             if (selection.getAnnotation().isSet()) {
                 // If an existing annotation was selected, take the feature editor model values from
                 // there
                 AnnotationFS annoFs = selectByAddr(aJCas, state.getSelection().getAnnotation()
                         .getId());
-                
+
                 // Try obtaining the layer from the feature structure
                 AnnotationLayer layer;
                 try {
@@ -1720,13 +922,13 @@ public class AnnotationDetailEditorPanel
                 if (!selection.isArc() && !state.getPreferences().isRememberLayer()) {
                     state.setSelectedAnnotationLayer(layer);
                 }
-                
+
                 loadFeatureEditorModelsCommon(aTarget, aJCas, layer, annoFs, null);
             }
             else {
                 // If a new annotation is being created, populate the feature editors from the
                 // remembered values (if any)
-                
+
                 if (selection.isArc()) {
                     // Avoid creation of arcs on locked layers
                     if (state.getSelectedAnnotationLayer() != null
@@ -1739,7 +941,7 @@ public class AnnotationDetailEditorPanel
                                 state.getRememberedArcFeatures());
                     }
                 }
-                else { 
+                else {
                     loadFeatureEditorModelsCommon(aTarget, aJCas,
                             state.getSelectedAnnotationLayer(), null,
                             state.getRememberedSpanFeatures());
@@ -1747,7 +949,7 @@ public class AnnotationDetailEditorPanel
             }
 
             annotationFeatureForm.updateRememberLayer();
-            
+
             if (aTarget != null) {
                 aTarget.add(annotationFeatureForm);
             }
@@ -1770,7 +972,7 @@ public class AnnotationDetailEditorPanel
             if (!feature.isEnabled()) {
                 continue;
             }
-            
+
             Serializable value = null;
             if (aFS != null) {
                 value = WebAnnoCasUtil.getFeature(aFS, feature);
@@ -1778,7 +980,7 @@ public class AnnotationDetailEditorPanel
             else if (aRemembered != null) {
                 value = aRemembered.get(feature);
             }
-            
+
             FeatureState featureState = null;
             if (WebAnnoConst.CHAIN_TYPE.equals(feature.getLayer().getType())) {
                 if (state.getSelection().isArc()) {
@@ -1798,10 +1000,10 @@ public class AnnotationDetailEditorPanel
             else {
                 featureState = new FeatureState(feature, value);
             }
-            
+
             if (featureState != null) {
                 state.getFeatureStates().add(featureState);
-                
+
                 // verification to check whether constraints exist for this project or NOT
                 if (state.getConstraints() != null && state.getSelection().getAnnotation().isSet()) {
                     // indicator.setRulesExist(true);
@@ -1844,7 +1046,7 @@ public class AnnotationDetailEditorPanel
             aAdapter.updateFeature(aJCas, featureState.feature,
                     state.getSelection().getAnnotation().getId(), featureState.value);
         }
-        
+
         // Generate info message
         if (state.getSelection().getAnnotation().isSet()) {
             String bratLabelText = TypeUtil.getUiLabelText(aAdapter,
@@ -1852,8 +1054,8 @@ public class AnnotationDetailEditorPanel
             info(generateMessage(state.getSelectedAnnotationLayer(), bratLabelText, false));
         }
     }
-        
-    
+
+
     protected void onChange(AjaxRequestTarget aTarget)
     {
         // Overriden in CurationPanel
@@ -1879,7 +1081,7 @@ public class AnnotationDetailEditorPanel
     {
         return (IModel<AnnotatorState>) getDefaultModel();
     }
-    
+
     public AnnotatorState getModelObject()
     {
         return (AnnotatorState) getDefaultModelObject();
@@ -1888,7 +1090,7 @@ public class AnnotationDetailEditorPanel
     /**
      * Clear the values from the feature editors.
      */
-    private void clearFeatureEditorModels(AjaxRequestTarget aTarget)
+    void clearFeatureEditorModels(AjaxRequestTarget aTarget)
     {
         LOG.trace("clearFeatureEditorModels()");
         getModelObject().getFeatureStates().clear();
@@ -1896,16 +1098,16 @@ public class AnnotationDetailEditorPanel
             aTarget.add(annotationFeatureForm);
         }
     }
-    
+
     /**
      * Adds and sorts tags based on Constraints rules
      */
     private void populateTagsBasedOnRules(JCas aJCas, FeatureState aModel)
     {
         LOG.trace("populateTagsBasedOnRules(feature: " + aModel.feature.getUiName() + ")");
-        
+
         AnnotatorState state = getModelObject();
-        
+
         // Add values from rules
         String restrictionFeaturePath;
         switch (aModel.feature.getLinkMode()) {
@@ -1921,9 +1123,9 @@ public class AnnotationDetailEditorPanel
                     + aModel.feature.getLinkMode() + "] on feature ["
                     + aModel.feature.getName() + "]");
         }
-        
+
         aModel.indicator.reset();
-        
+
         // Fetch possible values from the constraint rules
         List<PossibleValue> possibleValues;
         try {
@@ -1934,7 +1136,7 @@ public class AnnotationDetailEditorPanel
             //Only show indicator if this feature can be affected by Constraint rules!
             aModel.indicator.setAffected(evaluator.isThisAffectedByConstraintRules(
                     featureStructure, restrictionFeaturePath, state.getConstraints()));
-            
+
             possibleValues = evaluator.generatePossibleValues(
                     featureStructure, restrictionFeaturePath, state.getConstraints());
 
@@ -1949,7 +1151,7 @@ public class AnnotationDetailEditorPanel
 
         // Fetch actual tagset
         List<Tag> valuesFromTagset = annotationService.listTags(aModel.feature.getTagset());
-        
+
         // First add tags which are suggested by rules and exist in tagset
         List<Tag> tagset = compareSortAndAdd(possibleValues, valuesFromTagset, aModel.indicator);
 
@@ -1959,12 +1161,12 @@ public class AnnotationDetailEditorPanel
                 tagset.add(remainingTag);
             }
         }
-        
+
         // Record the possible values and the (re-ordered) tagset in the feature state
         aModel.possibleValues = possibleValues;
         aModel.tagset = tagset;
     }
-    
+
     /*
      * Compares existing tagset with possible values resulted from rule evaluation Adds only which
      * exist in tagset and is suggested by rules. The remaining values from tagset are added
@@ -1992,8 +1194,8 @@ public class AnnotationDetailEditorPanel
                     tag.setReordered(true);
                     // HACK END
                     //Avoid duplicate entries
-                    if(!returnList.contains(tag)){ 
-                        returnList.add(tag); 
+                    if(!returnList.contains(tag)){
+                        returnList.add(tag);
                     }
                 }
             }
@@ -2011,8 +1213,8 @@ public class AnnotationDetailEditorPanel
         state.getSelection().clear();
         clearFeatureEditorModels(aTarget);
     }
-    
-    private Set<AnnotationFS> getAttachedRels(JCas aJCas, AnnotationFS aFs, AnnotationLayer aLayer) {
+
+    Set<AnnotationFS> getAttachedRels(JCas aJCas, AnnotationFS aFs, AnnotationLayer aLayer) {
         Set<AnnotationFS> toBeDeleted = new HashSet<>();
         for (AnnotationLayer relationLayer : annotationService
                 .listAttachedRelationLayers(aLayer)) {
@@ -2035,7 +1237,7 @@ public class AnnotationDetailEditorPanel
                 relationTargetAttachFeature = targetFeature.getRange().getFeatureByBaseName(
                         relationAdapter.getAttachFeatureName());
             }
-            
+
             for (AnnotationFS relationFS : CasUtil.select(aJCas.getCas(), relationType)) {
                 // Here we get the annotations that the relation is pointing to in the UI
                 FeatureStructure sourceFS;
@@ -2063,10 +1265,10 @@ public class AnnotationDetailEditorPanel
                 }
             }
         }
-        
+
         return toBeDeleted;
     }
-    
+
     public AnnotationFeatureForm getAnnotationFeatureForm()
     {
         return annotationFeatureForm;
@@ -2074,18 +1276,18 @@ public class AnnotationDetailEditorPanel
 
     public Label getSelectedAnnotationLayer()
     {
-        return annotationFeatureForm.selectedAnnotationLayer;
+        return annotationFeatureForm.getSelectedAnnotationLayer();
     }
 
-    private boolean isForwardable()
+    boolean isForwardable()
     {
         AnnotatorState state = AnnotationDetailEditorPanel.this.getModelObject();
         AnnotationLayer selectedLayer = state.getSelectedAnnotationLayer();
-        
+
         if (selectedLayer == null) {
             return false;
         }
-        
+
 		if (selectedLayer.getId() <= 0) {
             return false;
         }
@@ -2093,32 +1295,32 @@ public class AnnotationDetailEditorPanel
 		if (!selectedLayer.getType().equals(WebAnnoConst.SPAN_TYPE)) {
             return false;
         }
-		
+
 		if (!selectedLayer.isLockToTokenOffset()) {
 			return false;
 		}
-		
+
 		// no forward annotation for multifeature layers.
         if (annotationService.listAnnotationFeature(selectedLayer).size() > 1) {
             return false;
         }
-		
+
         // if there are no features at all, no forward annotation
         if (annotationService.listAnnotationFeature(selectedLayer).isEmpty()) {
             return false;
         }
-        
+
         // we allow forward annotation only for a feature with a tagset
         if (annotationService.listAnnotationFeature(selectedLayer).get(0).getTagset() == null) {
             return false;
         }
-		
+
 		// there should be at least one tag in the tagset
         TagSet tagSet = annotationService.listAnnotationFeature(selectedLayer).get(0).getTagset();
 
         return annotationService.listTags(tagSet).size() != 0;
     }
-    
+
     public static void handleException(Component aComponent, AjaxRequestTarget aTarget,
             Exception aException)
     {
@@ -2143,7 +1345,7 @@ public class AnnotationDetailEditorPanel
             LOG.error("Error: " + e.getMessage(), e);
         }
     }
-    
+
     private static String generateMessage(AnnotationLayer aLayer, String aLabel, boolean aDeleted)
     {
         String action = aDeleted ? "deleted" : "created/updated";
@@ -2153,5 +1355,13 @@ public class AnnotationDetailEditorPanel
             msg += " Label: [" + aLabel + "]";
         }
         return msg;
+    }
+
+    AnnotationSchemaService getAnnotationService() {
+        return annotationService;
+    }
+
+    Logger getLog() {
+        return LOG;
     }
 }
